working script for jSON conversion
This is the script I was working on the last few days to get the output towards our format. It's a very preliminary script and not well annotated, but gets the job done for now. However, it only displays the variants as two arrays (one for names and one for confidence scores), so that's something you may want to change or adapt based on what suits you. Feel free to edit or use a different language to do it. You could also just import the jSONs produced into somewhere else and carry out the manipulations as you please, that's fully your choice. The script, the output table and example jSON are all in our DropBox folder (just change path and you should be able to run it). Let me know if any of it is unclear at all!